feat(profiles): Add constraint to unsername
Make unique
Make min 3 and max 50 (in the frontend we use only 20)
Add trigger function on insert and update to add short uuid
to username if we add an already existing name.
This is needed so we can use the first part of the email on signup.

The other functions are here because migra kept adding them.
Maybe it will finally shut up about them when I add them now